Changed LiveAgent snippet
Changed to an inline button which creates a pop-up.  Looks like the same code, so I'm not sure if it will work.  So far, no luck as I keep getting a GitHub email that says the commit succeeded but there's a problem with my DNS.  No idea what to do about that.